Make DominantResourceShare Compatible with Cohorts (#4097)
* Update DominantResourceShare to work with Cohorts

* Move DominantResourceShare methods to more appropriate files

* Cleanup resourceGroupNode, used only in tests

* Update documentation of several types
